Completed access control appwide.
Authorizations are now specified in the models for items, categories,
locations and users. These are enforced at a low level, so only admins
can create and edit locations, users, and categories; only logged in
users can submit listings, and only admins can submit listings under a
different username.

All views that link to restricted pages check for the user's permissions
with authenticates_access before displaying a link.

If a user happens upon an edit page for an item they are not authorized
to edit, they will get to see the form. However, it will fail to submit.
This will be fixed once we figure out a clean way to do it, but for now
no user can encounter that scenario without url hacking (or a bad
bookmark), and even still they can't do anything.

Also includes an update to authenticates_rpi to allow the 'sudo'
feature.